cleans auth and starts SAML support
