+++ composer.json
@@ -18,6 +18,7 @@
         "ext-intl": "*",
         "ext-mbstring": "*",
         "ext-pdo_mysql": "*",
+        "claroline/json-quiz": "dev-master",
         "composer/composer": "1.6.5",
         "doctrine/data-fixtures": "1.3.1",
         "doctrine/dbal": "2.7.1",
@@ -31,7 +32,7 @@
         "icap/html-diff": "1.1.0",
         "jdorn/sql-formatter": "1.2.17",
         "jms/security-extra-bundle": "1.6.1",
-        "claroline/json-quiz": "dev-master",
+        "lightsaml/sp-bundle": "1.2.1",
         "meyfa/php-svg": "0.8.0",
         "postal/postal": "1.0",
         "psr/log": "1.0.2",
