+++ main/authentication/Configuration/LegacyParametersMapping.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Claroline\AuthenticationBundle\Library;
+namespace Claroline\AuthenticationBundle\Configuration;
 
 use Claroline\AuthenticationBundle\Model\Oauth\OauthConfiguration;
 use Claroline\CoreBundle\Library\Configuration\LegacyParametersMappingInterface;
