+++ main/authentication/Resources/config/services.yml
@@ -1,7 +1,7 @@
 imports:
     - { resource: services/serializer.yml }
     - { resource: services/controller.yml }
-    - { resource: services/library.yml }
+    - { resource: services/configuration.yml }
     - { resource: services/manager.yml }
     - { resource: services/security.yml }
     - { resource: services/listener.yml }
