+++ main/authentication/Resources/config/suggested/lightsaml.yml
@@ -1,0 +1,15 @@
+light_saml_symfony_bridge:
+    own:
+        entity_id: http://localhost/lightsaml/demosp
+        credentials:
+            -
+                certificate: "%kernel.root_dir%/../vendor/lightsaml/lightsaml/web/sp/saml.crt"
+                key:         "%kernel.root_dir%/../vendor/lightsaml/lightsaml/web/sp/saml.key"
+                password:    ~
+    party:
+        idp:
+            files:
+                - "%kernel.root_dir%/../vendor/lightsaml/lightsaml/web/sp/openidp.feide.no.xml"
+                - "%kernel.root_dir%/../vendor/lightsaml/lightsaml/web/sp/testshib-providers.xml"
+    store:
+        id_state: id_store  # name of id store service created in step 6
